XmlFSocket: protocol change
The header no more contains the size of the XML block. This makes
sending easier, as one does not need to first construct the entire
document and then afterward inject the resulted size in the header.

A footer tag is now used in addition of the header for making it trivial
to detect when a full xml document has been received.