Debugger setup for inspecting different actor compositions. Position snapshot&recall reverted to working state.
Add a breakpoint to RagdollControllerGameMode.cpp: line with comment 'add a breakpoint here'.
This enumerates actors with different ways to have a StaticMeshComponent, and enumerates the
components of these actors.

Some of the actors are StaticMeshActors, some are StaticMeshActors converted to blueprints, and some
are pure blueprint actors (non-derived AActor) with a StaticMeshComponent.

Result: all contain a StaticMeshComponent in their component array. The StaticMeshActors
additionally contain a pointer as a member field that points to that same component.

See also https://trello.com/c/kGHuH99j

Also, reverted back the position vector snapshot and recall test code that already worked in a
previous commit.